moved TGD to com.revature.util
